// src/client/models/place.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Coord } from './coord';
// May contain unused imports in some cases
// @ts-ignore
import type { Geocode } from './geocode';
// May contain unused imports in some cases
// @ts-ignore
import type { PlaceAdministrativeRegion } from './place-administrative-region';
// May contain unused imports in some cases
// @ts-ignore
import type { PlaceAdministrativeRegionsInner } from './place-administrative-regions-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { PlacePoi } from './place-poi';
// May contain unused imports in some cases
// @ts-ignore
import type { StopArea } from './stop-area';
// May contain unused imports in some cases
// @ts-ignore
import type { StopPoint } from './stop-point';

export interface Place {
    /**
     * Place ID
     */
    'id'?: string;
    /**
     * Place name
     */
    'name'?: string;
    /**
     * Quality score (0-100, higher is better)
     */
    'quality'?: number;
    /**
     * Embedded type - determines which embedded object is populated
     */
    'embedded_type'?: PlaceEmbeddedTypeEnum;
    'stop_area'?: StopArea;
    'stop_point'?: StopPoint;
    'address'?: Geocode;
    'poi'?: PlacePoi;
    'administrative_region'?: PlaceAdministrativeRegion;
    'coord'?: Coord;
    /**
     * List of administrative regions this place belongs to
     */
    'administrative_regions'?: Array<PlaceAdministrativeRegionsInner>;
    /**
     * Distance in meters (for places_nearby endpoint)
     */
    'distance'?: number;
}

export const PlaceEmbeddedTypeEnum = {
    StopArea: 'stop_area',
    StopPoint: 'stop_point',
    Address: 'address',
    Poi: 'poi',
    AdministrativeRegion: 'administrative_region'
} as const;

export type PlaceEmbeddedTypeEnum = typeof PlaceEmbeddedTypeEnum[keyof typeof PlaceEmbeddedTypeEnum];