// src/client/models/stop-time.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StopPoint } from './stop-point';

export interface StopTime {
    /**
     * Arrival time (HHmmss)
     */
    'arrival_time'?: string;
    /**
     * Departure time (HHmmss)
     */
    'departure_time'?: string;
    /**
     * Headsign
     */
    'headsign'?: string;
    /**
     * Whether pickup is allowed
     */
    'pickup_allowed'?: boolean;
    /**
     * Whether drop off is allowed
     */
    'drop_off_allowed'?: boolean;
    /**
     * Whether stop is skipped
     */
    'skipped_stop'?: boolean;
    'stop_point'?: StopPoint;
    /**
     * UTC arrival time (YYYYMMDDTHHmmss)
     */
    'utc_arrival_time'?: string;
    /**
     * UTC departure time (YYYYMMDDTHHmmss)
     */
    'utc_departure_time'?: string;
}